# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /charts/{id}/regenerate:
    post:
      operationId: regenerateChart
      tags: [charts]
      summary: Regenerate the spec for a single chart
      description: |
        Re-runs the research pipeline for just this chart, replacing its spec
        in place. Requires a DRD uploaded for the CE. Locked CEs return 409.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Regenerated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartWithCe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: CE is locked from automated regeneration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "412":
          description: No DRD uploaded for the CE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Pipeline failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /charts/{id}/feedback:
    get:
      operationId: listChartFeedback
      tags: [feedback]
      summary: List feedback for a single chart
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feedback list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartFeedback"
    post:
      operationId: createChartFeedback
      tags: [feedback]
      summary: Submit feedback on a chart
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChartFeedbackInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartFeedback"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Chart not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /feedback:
    get:
      operationId: listAllFeedback
      tags: [feedback]
      summary: List feedback across all charts (for triage)
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Triage feedback rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TriageFeedback"
  /feedback/{id}:
    patch:
      operationId: updateFeedback
      tags: [feedback]
      summary: Update a feedback row's status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartFeedback"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Ce:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        city:
          type: string
        country:
          type: string
        category:
          type: string
        summary:
          type: string
        emoji:
          type: string
        status:
          type: string
        chartCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - name
        - city
        - country
        - category
        - summary
        - emoji
        - status
        - chartCount
        - createdAt
        - updatedAt
    Chart:
      type: object
      properties:
        id:
          type: integer
        ceId:
          type: integer
        slug:
          type: string
        question:
          type: string
        title:
          type: string
        subtitle:
          type: string
        insight:
          type: string
        chartType:
          type: string
        spec:
          type: object
          additionalProperties: true
        status:
          type: string
        provenance:
          type: object
          nullable: true
          additionalProperties: true
        sortOrder:
          type: integer
        openFeedbackCount:
          type: integer
        topFeedbackSeverity:
          type: string
          nullable: true
        editCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - ceId
        - slug
        - question
        - title
        - subtitle
        - insight
        - chartType
        - spec
        - sortOrder
        - createdAt
        - updatedAt
    ChartWithCe:
      type: object
      properties:
        chart:
          $ref: "#/components/schemas/Chart"
        ce:
          $ref: "#/components/schemas/Ce"
      required:
        - chart
        - ce
...
    ChartFeedback:
      type: object
      properties:
        id:
          type: integer
        chartId:
          type: integer
        rating:
          type: integer
          nullable: true
        issueCategory:
          type: string
          nullable: true
        note:
          type: string
        reporterName:
          type: string
        status:
          type: string
        severity:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - chartId
        - note
        - reporterName
        - status
        - severity
        - createdAt
        - updatedAt
    ChartFeedbackInput:
      type: object
      properties:
        rating:
          type: integer
          nullable: true
        issueCategory:
          type: string
          nullable: true
        note:
          type: string
        reporterName:
          type: string
...
    FeedbackUpdate:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    TriageFeedback:
      type: object
      properties:
        feedback:
          $ref: "#/components/schemas/ChartFeedback"
        chart:
          $ref: "#/components/schemas/Chart"
        ce:
          $ref: "#/components/schemas/Ce"
        chartEditCount:
          type: integer
      required:
        - feedback
        - chart
        - ce
        - chartEditCount
...