# lib/api-zod/src/generated/types/chartFeedbackInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChartFeedbackInput {
  rating?: number | null;
  issueCategory?: string | null;
  note?: string;
  reporterName?: string;
}