# lib/api-zod/src/generated/types/triageFeedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Ce } from "./ce";
import type { Chart } from "./chart";
import type { ChartFeedback } from "./chartFeedback";

export interface TriageFeedback {
  feedback: ChartFeedback;
  chart: Chart;
  ce: Ce;
  chartEditCount: number;
}